Struct postflop_solver::ActionTree

pub struct ActionTree { /* private fields */ }

A struct representing an abstract game tree.

Implementations

impl ActionTree

pub fn new(config: TreeConfig) -> Result<Self, String>

Creates a new ActionTree with the specified configuration.

pub fn config(&self) -> &TreeConfig

Obtains the configuration of the game tree.

pub fn added_lines(&self) -> &[Vec<Action>]

Obtains the list of added lines.

pub fn removed_lines(&self) -> &[Vec<Action>]

Obtains the list of removed lines.

pub fn invalid_terminals(&self) -> Vec<Vec<Action>>

Returns a list of all terminal nodes that should not be.

pub fn add_line(&mut self, line: &[Action]) -> Result<(), String>

Adds a given line to the action tree.

line except the last action must exist in the current tree.

The last action of the line must not exist in the current tree.

Except for the case that the line is in the removed lines, the last action of the line must be a bet action (including raise and all-in action).

Chance actions (i.e., dealing turn and river cards) must be omitted from the line.

pub fn remove_line(&mut self, line: &[Action]) -> Result<(), String>

Removes a given line from the action tree.

line must exist in the current tree.

Chance actions (i.e., dealing turn and river cards) must be omitted from the line.

If the current node is removed by this method, the current node is moved to the nearest ancestor node that is not removed.

pub fn back_to_root(&mut self)

Moves back to the root node.

pub fn history(&self) -> &[Action]

Obtains the current action history.

pub fn apply_history(&mut self, history: &[Action]) -> Result<(), String>

Applies the given action history from the root node.

pub fn is_terminal_node(&self) -> bool

Returns whether the current node is a terminal node.

pub fn is_chance_node(&self) -> bool

Returns whether the current node is a chance node.

pub fn available_actions(&self) -> &[Action]

Returns the available actions for the current node.

If the current node is a chance node, returns possible actions after the chance event.

pub fn play(&mut self, action: Action) -> Result<(), String>

Plays the given action. Returns Ok(()) if the action is valid.

The action must be one of the possible actions at the current node. If the current node is a chance node, the chance action is automatically played before playing the given action.

pub fn undo(&mut self) -> Result<(), String>

Undoes the last action. Returns Ok(()) if the action is successfully undone.

pub fn add_action(&mut self, action: Action) -> Result<(), String>

Adds a given action to the current node.

Internally, this method calls add_line with the current action history and the given action. See add_line for the details.

pub fn remove_action(&mut self, action: Action) -> Result<(), String>

Removes a given action from the current node.

Internally, this method calls remove_line with the current action history and the given action. See remove_line for the details.

pub fn remove_current_node(&mut self) -> Result<(), String>

Removes the current node.

Internally, this method calls remove_line with the current action history. See remove_line for the details.

pub fn total_bet_amount(&self) -> [i32; 2]

Returns the total bet amount of each player (OOP, IP).