Struct postflop_solver::GameConfig

pub struct GameConfig {

    pub flop: [u8; 3],
    pub turn: u8,
    pub river: u8,
    pub starting_pot: i32,
    pub effective_stack: i32,
    pub range: [Range; 2],
    pub flop_bet_sizes: [BetSizeCandidates; 2],
    pub turn_bet_sizes: [BetSizeCandidates; 2],
    pub river_bet_sizes: [BetSizeCandidates; 2],
    pub turn_donk_sizes: Option<DonkSizeCandidates>,
    pub river_donk_sizes: Option<DonkSizeCandidates>,
    pub add_all_in_threshold: f32,
    pub force_all_in_threshold: f32,
    pub adjust_bet_size_before_all_in: bool,

}

A struct for postflop game configuration.

Examples

use postflop_solver::*;

let oop_range = "66+,A8s+,A5s-A4s,AJo+,K9s+,KQo,QTs+,JTs,96s+,85s+,75s+,65s,54s";
let ip_range = "QQ-22,AQs-A2s,ATo+,K5s+,KJo+,Q8s+,J8s+,T7s+,96s+,86s+,75s+,64s+,53s+";
let bet_sizes = BetSizeCandidates::try_from(("60%, e, a", "2.5x")).unwrap();

let config = GameConfig {
    flop: flop_from_str("Td9d6h").unwrap(),
    turn: card_from_str("Qc").unwrap(),
    river: NOT_DEALT,
    starting_pot: 200,
    effective_stack: 900,
    range: [oop_range.parse().unwrap(), ip_range.parse().unwrap()],
    flop_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    turn_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    river_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    turn_donk_sizes: None,
    river_donk_sizes: Some(DonkSizeCandidates::try_from("50%").unwrap()),
    add_all_in_threshold: 1.2,
    force_all_in_threshold: 0.1,
    adjust_bet_size_before_all_in: false,
};

Fields
flop: [u8; 3]

Flop cards: each card must be unique and in range [0, 52).

turn: u8

Turn card: must be in range [0, 52) or NOT_DEALT.

river: u8

River card: must be in range [0, 52) or NOT_DEALT.

starting_pot: i32

Initial pot size: must be greater than 0.

effective_stack: i32

Initial effective stack size: must be greater than 0.

range: [Range; 2]

Initial range of each player.

flop_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player for the flop.

turn_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player for the turn.

river_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player for the river.

turn_donk_sizes: Option<DonkSizeCandidates>

Donk size candidates for the turn (set None to use default sizes).

river_donk_sizes: Option<DonkSizeCandidates>

Donk size candidates for the river (set None to use default sizes).

add_all_in_threshold: f32

Add all-in action if the current SPR is below this value (set 0.0 to disable).

force_all_in_threshold: f32

Force all-in action if the SPR after the opponent’s call is below this value (set 0.0 to disable).

adjust_bet_size_before_all_in: bool

Enable bet size adjustment for the bet just before the opponent’s all-in.