Struct postflop_solver::Range

pub struct Range { /* private fields */ }

A struct representing a player’s range.

The Range struct implements the FromStr trait, so you can construct a range from a string using parse::<Range>(). The string must be in the following format (similar to PioSOLVER):

Each group is separated by a comma. (e.g., “AA,AKs”)

Each group can have an optional weight separated by a colon. (e.g., “AA:0.5”)

Each group must be one of the following:

Singleton (e.g., “AA”, “AKs”, “AKo”, “AsAh”)

Plus range (e.g., “TT+”, “ATs+”, “T9o+”)

Dash range (e.g., “QQ-88”, “A9s-A6s”, “98o-65o”)

Examples

use postflop_solver::Range;

// construct a range from a string
let range = "QQ+,AKs".parse::<Range>().unwrap();

// rank is defined as follows: A => 12, K => 11, ..., 2 => 0
let ace_rank = 12;
let king_rank = 11;
let queen_rank = 10;

// check that the hand "QQ" is in the range
assert_eq!(range.get_weight_pair(queen_rank), 1.0);

// check that the hand "AKo" is not in the range
assert_eq!(range.get_weight_offsuit(ace_rank, king_rank), 0.0);

Implementations

impl Range

pub fn new() -> Self

Creates an empty range.

pub fn ones() -> Self

Creates a full range.

pub fn from_raw_data(data: &[f32]) -> Result<Self, String>

Attempts to create a range from raw data.

pub fn raw_data(&self) -> &[f32]

Obtains the raw data of the range.

pub fn from_hands_weights( hands: &[(Card, Card)], weights: &[f32] ) -> Result<Self, String>

Attempts to create a range from a list of hands with their weights.

pub fn get_hands_weights( &self, dead_cards_mask: u64 ) -> (Vec<(Card, Card)>, Vec<f32>)

Returns a list of all hands in this range and their associated weights.

If there are no dead cards, pass 0 to dead_cards_mask. The returned hands are sorted in lexicographical order.

pub fn from_sanitized_str(ranges: &str) -> Result<Self, String>

Attempts to create a range from a sanitized range string.

“Sanitized” means that the range string does not contain any invalid patterns and whitespace characters. Therefore, this method can bypass the regular expression processing. If you want to create a range from a regular string, use parse::<Range>() instead.

pub fn clear(&mut self)

Clears the range.

pub fn is_empty(&self) -> bool

Returns whether the range is empty.

pub fn invert(&mut self)

Inverts the range.

pub fn get_weight_by_cards(&self, card1: Card, card2: Card) -> f32

Obtains the weight of a specified hand.

Undefined behavior if:

card1 or card2 is not less than 52

card1 is equal to card2

pub fn get_weight_pair(&self, rank: u8) -> f32

Obtains the average weight of specified pair hands.

Undefined behavior if rank is not less than 13.

pub fn get_weight_suited(&self, rank1: u8, rank2: u8) -> f32

Obtains the average weight of specified suited hands.

Undefined behavior if:

rank1 or rank2 is not less than 13

rank1 is equal to rank2

pub fn get_weight_offsuit(&self, rank1: u8, rank2: u8) -> f32

Obtains the average weight of specified offsuit hands.

Undefined behavior if:

rank1 or rank2 is not less than 13

rank1 is equal to rank2

pub fn set_weight_by_cards(&mut self, card1: Card, card2: Card, weight: f32)

Sets the weight of a specified hand.

Undefined behavior if:

card1 or card2 is not less than 52

card1 is equal to card2

weight is not in the range [0.0, 1.0]

pub fn set_weight_pair(&mut self, rank: u8, weight: f32)

Sets the weights of specified pair hands.

Undefined behavior if:

rank is not less than 13

weight is not in the range [0.0, 1.0]

pub fn set_weight_suited(&mut self, rank1: u8, rank2: u8, weight: f32)

Sets the weights of specified suited hands.

Undefined behavior if:

rank1 or rank2 is not less than 13

rank1 is equal to rank2

weight is not in the range [0.0, 1.0]

pub fn set_weight_offsuit(&mut self, rank1: u8, rank2: u8, weight: f32)

Sets the weights of specified offsuit hands.

Undefined behavior if:

rank1 or rank2 is not less than 13

rank1 is equal to rank2

weight is not in the range [0.0, 1.0]