Struct postflop_solver::GameConfig

pub struct GameConfig {
    pub flop: [u8; 3],
    pub turn: u8,
    pub river: u8,
    pub starting_pot: i32,
    pub effective_stack: i32,
    pub range: [Range; 2],
    pub flop_bet_sizes: [BetSizeCandidates; 2],
    pub turn_bet_sizes: [BetSizeCandidates; 2],
    pub river_bet_sizes: [BetSizeCandidates; 2],
    pub add_all_in_threshold: f32,
    pub force_all_in_threshold: f32,
    pub adjust_last_two_bet_sizes: bool,
}

A struct for postflop game configuration.

Examples

use postflop_solver::*;

let oop_range = "66+,A8s+,A5s-A4s,AJo+,K9s+,KQo,QTs+,JTs,96s+,85s+,75s+,65s,54s";
let ip_range = "QQ-22,AQs-A2s,ATo+,K5s+,KJo+,Q8s+,J8s+,T7s+,96s+,86s+,75s+,64s+,53s+";
let bet_sizes = BetSizeCandidates::try_from(("50%", "50%")).unwrap();

let config = GameConfig {
    flop: flop_from_str("Td9d6h").unwrap(),
    turn: NOT_DEALT, // or `card_from_str("As").unwrap()`
    river: NOT_DEALT,
    starting_pot: 200,
    effective_stack: 900,
    range: [oop_range.parse().unwrap(), ip_range.parse().unwrap()],
    flop_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    turn_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    river_bet_sizes: [bet_sizes.clone(), bet_sizes.clone()],
    add_all_in_threshold: 1.2,
    force_all_in_threshold: 0.1,
    adjust_last_two_bet_sizes: true,
};

Fields
flop: [u8; 3]

Flop cards: each card must be unique and in range [0, 52).

turn: u8

Turn card: must be in range [0, 52) or NOT_DEALT.

river: u8

River card: must be in range [0, 52) or NOT_DEALT.

starting_pot: i32

Initial pot size: must be greater than 0.

effective_stack: i32

Initial effective stack size: must be greater than 0.

range: [Range; 2]

Initial range of each player.

flop_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player in flop.

turn_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player in turn.

river_bet_sizes: [BetSizeCandidates; 2]

Bet size candidates of each player in river.

add_all_in_threshold: f32

Add all-in action when SPR is below this value (set 0.0 to disable).

force_all_in_threshold: f32

Force all-in action when the ratio of opponent’s next raise size to the pot size will be less than this value (set 0.0 to disable).

adjust_last_two_bet_sizes: bool

Enable bet size adjustment of last two bets.