Struct postflop_solver::PostFlopGame

pub struct PostFlopGame { /* private fields */ }

A struct representing a postflop game.

Implementations

impl PostFlopGame

pub fn new() -> Self

Constructs a new empty PostFlopGame (needs update_config() before solving).

pub fn with_config(config: &GameConfig) -> Result<Self, String>

Constructs a new PostFlopGame instance with the given configuration.

pub fn update_config(&mut self, config: &GameConfig) -> Result<(), String>

Updates the game configuration. The solved result will be lost.

pub fn config(&self) -> &GameConfig

Obtains the game configuration.

pub fn private_hand_cards(&self, player: usize) -> &[(u8, u8)]

Returns the card list of private hands of the given player.

pub fn memory_usage(&self) -> (u64, u64)

Returns the estimated memory usage in bytes (uncompressed, compressed).

pub fn allocate_memory(&mut self, enable_compression: bool)

Allocates the memory.

pub fn back_to_root(&mut self)

Moves the current node back to the root node.

pub fn available_actions(&self) -> Vec<Action>

Returns the available actions for the current node.

Note: If the current node is a chance node, isomorphic chances are grouped together into one representative action.

pub fn is_terminal_action(&self) -> Vec<bool>

Returns whether the available actions are terminal.

Note that the call action after the all-in action is considered as terminal.

pub fn is_chance_node(&self) -> bool

Returns whether the current node is a chance node.

pub fn possible_cards(&self) -> u64

If the current node is a chance node, returns a list of cards that may be dealt.

The returned value is a 64-bit integer. The i-th bit is set to 1 if the card of ID i may be dealt. If the current node is not a chance node, returns 0.

Card ID: "2c" => 0, "2d" => 1, "2h" => 2, …, "As" => 51.

pub fn current_player(&self) -> usize

Returns the current player (0 = OOP, 1 = IP).

If the current node is a chance node, returns an undefined value.

pub fn play(&mut self, action: usize)

Plays the given action. Playing a terminal action is not allowed.

action

If the current node is a chance node, action corresponds to the card ID of the dealt card.

If the current node is not a chance node, plays the action-th action of available_actions().

pub fn cache_normalized_weights(&mut self)

Computes the normalized weights and caches them.

After mutating the current node, this method must be called before calling normalized_weights(), expected_values(), or equity().

pub fn weights(&self, player: usize) -> &[f32]

Returns the weights of each private hand of the given player.

pub fn normalized_weights(&self, player: usize) -> &[f64]

Returns the normalized weights of each private hand of the given player.

The “normalized weights” represent the actual number of combinations that the player is holding each hand.

After mutating the current node, you must call the cache_normalized_weights() method before calling this method.

pub fn expected_values(&self) -> Vec<f32>

Returns the expected values of each private hand of the current player.

Panics if the current node is a chance node.

After mutating the current node, you must call the cache_normalized_weights() method before calling this method.

pub fn equity(&self) -> Vec<f32>

Returns the equity of each private hand of the current player.

Panics if the current node is a chance node.

After mutating the current node, you must call the cache_normalized_weights() method before calling this method.

pub fn strategy(&self) -> Vec<f32>

Returns the strategy of the current player.

Panics if the current node is a chance node.

The strategy is a vector of the length of #(actions) * #(private hands). The probability of i-th action with j-th private hand is stored in the i * #(private hands) + j-th element.