Struct postflop_solver::PostFlopGame

pub struct PostFlopGame { /* private fields */ }

A struct representing a postflop game.

Implementations

impl PostFlopGame

pub fn new() -> Self

Creates a new empty PostFlopGame (needs update_config() before solving).

pub fn with_config(
card_config: CardConfig,
action_tree: ActionTree
) -> Result<Self, String>

Creates a new PostFlopGame with the specified configuration.

pub fn update_config(
&mut self,
card_config: CardConfig,
action_tree: ActionTree
) -> Result<(), String>

Updates the game configuration. The solved result will be lost.

pub fn card_config(&self) -> &CardConfig

Obtains the card configuration.

pub fn tree_config(&self) -> &TreeConfig

Obtains the tree configuration.

pub fn added_lines(&self) -> &[Vec<Action>]

Obtains the added lines.

pub fn removed_lines(&self) -> &[Vec<Action>]

Obtains the removed lines.

pub fn private_cards(&self, player: usize) -> &[(u8, u8)]

Returns the card list of private hands of the given player.

pub fn memory_usage(&self) -> (u64, u64)

Returns the estimated memory usage in bytes (uncompressed, compressed).

pub fn allocate_memory(&mut self, enable_compression: bool)

Allocates the memory.

pub fn back_to_root(&mut self)

Moves the current node back to the root node.

pub fn history(&self) -> &[usize]

Returns the history of the current node.

The history is a list of action indices, i.e., the arguments of play. If usize::MAX was passed to play, it is replaced with the actual action index.

pub fn apply_history(&mut self, history: &[usize])

Applies the given history from the root node.

This method first calls back_to_root and then calls play for each action in the history. The action of usize::MAX is allowed for chance nodes.

pub fn is_terminal_node(&self) -> bool

Returns whether the current node is a terminal node.

Note that the turn/river node after the call action after the all-in action is considered terminal.

pub fn is_chance_node(&self) -> bool

Returns whether the current node is a chance node (i.e., turn/river node).

Note that the terminal node is not considered a chance node.

pub fn available_actions(&self) -> &[Action]

Returns the available actions for the current node.

If the current node is a terminal, returns an empty list. If the current node is a turn/river node and not a terminal, isomorphic chances are grouped into one representative action (in most cases, you should use the possible_cards method).

pub fn possible_cards(&self) -> u64

If the current node is a chance node, returns a list of cards that may be dealt.

The returned value is a 64-bit integer. The i-th bit is set to 1 if the card of ID i may be dealt. If the current node is not a chance node, returns 0.

Card ID: "2c" => 0, "2d" => 1, "2h" => 2, …, "As" => 51.

pub fn current_player(&self) -> usize

Returns the current player (0 = OOP, 1 = IP).

If the current node is a terminal node or a chance node, returns an undefined value.

pub fn current_board(&self) -> Vec<u8>

Returns the current board.

The returned vector is of length 3, 4, or 5. The flop cards, the turn card, and the river card, if any, are stored in this order.

pub fn play(&mut self, action: usize)

Plays the given action. Playing an action from a terminal node is not allowed.

action

If the current node is a chance node, the action corresponds to the card ID of the dealt card. The action can be usize::MAX, in which case the actual card is chosen from possible cards.

If the current node is not a chance node, plays the action-th action of available_actions.

Panics if the memory is not yet allocated or the current node is a terminal node.

pub fn cache_normalized_weights(&mut self)

Computes the normalized weights and caches them.

After mutating the current node, this method must be called once before calling normalized_weights, equity, expected_values, or expected_values_detail.

pub fn weights(&self, player: usize) -> &[f32]

Returns the weights of each private hand of the given player.

If a hand overlaps with the board, an undefined value is returned.

pub fn normalized_weights(&self, player: usize) -> &[f32]

Returns the normalized weights of each private hand of the given player.

The “normalized weights” represent the actual number of combinations that the player is holding each hand.

After mutating the current node, you must call the cache_normalized_weights method before calling this method.

pub fn equity(&self, player: usize) -> Vec<f32>

Returns the equity of each private hand of the given player.

After mutating the current node, you must call the cache_normalized_weights method before calling this method.

pub fn expected_values(&self, player: usize) -> Vec<f32>

Returns the expected values of each private hand of the given player.

Panics if the game is not solved.

After mutating the current node, you must call the cache_normalized_weights method before calling this method.

pub fn expected_values_detail(&self, player: usize) -> Vec<f32>

Returns the expected values of each action of each private hand of the given player.

If the given player is the current player, the return value is a vector of the length of #(actions) * #(private hands). The expected value of i-th action with j-th private hand is stored in the i * #(private hands) + j-th element.

Otherwise, this method is the same as the expected_values method, so the return vector is the length of #(private hands).

Panics if the game is not solved.

After mutating the current node, you must call the cache_normalized_weights method before calling this method.

pub fn strategy(&self) -> Vec<f32>

Returns the strategy of the current player.

The return value is a vector of the length of #(actions) * #(private hands). The probability of i-th action with j-th private hand is stored in the i * #(private hands) + j-th element.

If a hand overlaps with the board, an undefined value is returned.

Panics if the current node is a terminal node or a chance node. Also, panics if the memory is not yet allocated.

pub fn total_bet_amount(&self) -> [i32; 2]

Returns the total bet amount of each player (OOP, IP).